Fix issues with ws switch and roslib refactor
