+++ src/lib/controller/gamepad.ts
@@ -1,5 +1,5 @@
 import ControllerState from './controllerState';
-import { useControllerStore } from '@/lib/store/controller';
+import { useControllerStore } from '@/store/controllerStore';
 
 const DELTA_SENSITIVITY = 0.01;
 const POLLING_RATE_IN_HERTZ = 20;
