+++ src/store/exampleStore.ts
